create new component for visuals and new component for chart
