+++ web-app/client/components/AnalyticsChart.js
@@ -1,0 +1,1 @@
+/* Eventually where descriptive statistics chart will go */
